@@ -36,10 +36,9 @@
     "path": "^0.12.7",
     "postcss": "^8.4.6",
     "ramda": "^0.28.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
     "react-hook-form": "^7.28.1",
-    "react-query": "^3.34.13",
     "react-router-dom": "^6.2.2",
     "react-spinners": "^0.11.0",
     "rxjs": "^7.5.5",
@@ -52,8 +51,8 @@
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
     "@types/jest": "^27.4.0",
-    "@types/react": "^17.0.33",
-    "@types/react-dom": "^17.0.10",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
     "@types/react-test-renderer": "^17.0.1",
     "@vitejs/plugin-react": "^1.0.7",
     "effector-logger": "^0.13.4",
